% Generated by roxygen2 (4.0.2): do not edit by hand
\name{animatedGIF}
\alias{animatedGIF}
\title{Create an animated gif to be used for embedding in a Markdown file}
\usage{
animatedGIF(expr, filename, title = filename, allowHTML = TRUE, ...)
}
\value{
a Markdown/HTML snippet that can be used to embed the generated animation
}
\description{
Create an animated gif to be used for embedding in a Markdown file
}

