% Generated by roxygen2 (4.0.2): do not edit by hand
\name{animatedHTML}
\alias{animatedHTML}
\title{Creates an HTML file for displaying and navigating an animation}
\usage{
animatedHTML(expr, filename, title = filename,
  ani.width = ani.options("ani.width"),
  ani.height = ani.options("ani.height"), ...)
}
\value{
an HTML-Iframe snippet that can be used to embed the generated animation
}
\description{
Creates an HTML file for displaying and navigating an animation
}

