% Generated by roxygen2 (4.0.2): do not edit by hand
\name{animatedHTML}
\alias{animatedHTML}
\title{Create an HTML file for displaying and navigating an animation}
\usage{
animatedHTML(expr, filename, title = filename,
  ani.width = ani.options("ani.width"),
  ani.height = ani.options("ani.height"), ...)
}
\arguments{
\item{expr}{an R expression to be evaluated to create a sequence of images}

\item{filename}{unique filename to be used for the saved image files (without extension)}

\item{title}{animation title, used for the animation page as well as Alt text for the Iframe}

\item{...}{other arguments passed to \code{\link{ani.options}}}
}
\value{
an HTML-Iframe snippet that can be used to embed the generated animation
}
\description{
Create an HTML file for displaying and navigating an animation
}
\seealso{
\code{\link{saveHTML}}
}

